requirements.txt
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
langchain>=0.1.0
langgraph>=0.0.20
langchain-openrouter>=0.1.0
fastapi>=0.100.0
python-multipart>=0.0.6
uvicorn>=0.20.0
